document.c: pass NULL as the 4th argument to pdf_save_document()
Passing a `{ 0 }' as 4th argument to pdf_save_document() is no longer needed.
Just pass NULL and get rid of opts.

At least in recent mupdf (version 1.9a) `*in_opts' (the 4th argument of
pdf_save_document()) is checked if NULL and in that case the default options are
used.

Signed-off-by: Sebastian Ramacher <sebastian@ramacher.at>