fix: pdf metadata is not shown
The command `:info` does not display the metadata of the PDF document.
Instead, a white line is shown at the bottom of the screen and an error
message is logged to the console. This is because of a misinterpretation
of the string buffer, returned from `pdf_to_str_buf`. The content of the
string buffer is `UTF-16be` encoded and therefore not suitable for
passing to the zathura API, which expects an `UTF-8` encoded sequence.
By using the function `pdf_to_text_string` the `UTF-8` conversion is
done by `mupdf` and no further steps are required.